Change _serialize
