+++ aiaccel/optimizer/nelder_mead_optimizer.py
@@ -336,48 +336,3 @@ class NelderMeadOptimizer(AbstractOptimizer):
                 self.parameter_pool.append(copy.copy(p))
 
         self._generate_hp_ready(number)
-
-    def _serialize(self, trial_id) -> None:
-        """Serialize this module.
-        Returns:
-            dict: The serialized objects.
-        """
-
-        parameter_pool = copy.deepcopy(self.parameter_pool)
-        for p_pool in parameter_pool:
-            for p_pool_param in p_pool['parameters']:
-                if type(p_pool_param['value']) is np.float64:
-                    p_pool_param['value'] = float(p_pool_param['value'])
-
-        self.storage.variable.d['native_random_state'].set(trial_id, self.get_native_random_state())
-        self.storage.variable.d['numpy_random_state'].set(trial_id, self.get_numpy_random_state())
-        self.storage.variable.d['num_of_generated_parameter'].set(trial_id, self.num_of_generated_parameter)
-        self.storage.variable.d['loop_count'].set(trial_id, self.loop_count)
-        self.storage.variable.d['parameter_pool'].set(trial_id, parameter_pool)
-        self.storage.variable.d['nelder_mead'].set(trial_id, self.nelder_mead.serialize())
-        self.storage.variable.d['order'].set(trial_id, self.order)
-
-    def _deserialize(self, trial_id: int) -> None:
-        """Deserialize this module.
-        Args:
-            dict_objects(dict): A dictionary including serialized objects.
-        Returns:
-            None
-        """
-        self.set_native_random_state(self.storage.variable.d['native_random_state'].get(trial_id))
-        self.set_numpy_random_state(self.storage.variable.d['numpy_random_state'].get(trial_id))
-
-        self.num_of_generated_parameter = self.storage.variable.d['num_of_generated_parameter'].get(trial_id)
-        self.loop_count = self.storage.variable.d['loop_count'].get(trial_id)
-
-        parameter_pool = copy.deepcopy(self.storage.variable.d['parameter_pool'].get(trial_id))
-        for p_pool in parameter_pool:
-            for p_pool_param in p_pool['parameters']:
-                if type(p_pool_param['value']) is float:
-                    p_pool_param['value'] = np.float64(p_pool_param['value'])
-        self.parameter_pool = parameter_pool
-
-        self.nelder_mead = NelderMead(self.params.get_parameter_list())
-        self.nelder_mead.deserialize(self.storage.variable.d['nelder_mead'].get(trial_id))
-
-        self.order = self.storage.variable.d['order'].get(trial_id)
