+++ aiaccel/scheduler/abstract_scheduler.py
@@ -255,9 +255,9 @@ class AbstractScheduler(AbstractModule):
         return True
 
     def _serialize(self, trial_id) -> None:
-        self.storage.variable.d['native_random_state'].set(trial_id, self.get_native_random_state())
-        self.storage.variable.d['numpy_random_state'].set(trial_id, self.get_numpy_random_state())
-        self.storage.variable.d['loop_count'].set(trial_id, self.loop_count)
+        self.storage.variable.d['native_random_state'].set(trial_id=trial_id, value=self.get_native_random_state())
+        self.storage.variable.d['numpy_random_state'].set(trial_id=trial_id, value=self.get_numpy_random_state())
+        self.storage.variable.d['loop_count'].set(trial_id=trial_id, value=self.loop_count)
 
     def _deserialize(self, trial_id: int) -> None:
         """Deserialize this module.
