+++ tests/resumption/resumption_test.py
@@ -36,4 +36,7 @@ class ResumptionTest(IntegrationTest):
 
     def get_final_result(self, storage):
         data = storage.result.get_all_result()
+        print([d.objective for d in data])
         return [d.objective for d in data][-1]
+
+
